refactor sink handling into setup lib
